Rename streams to async_streams.
