+++ src/contrib/data/async_streams/byte_stream.ts
