+++ src/contrib/data/async_streams/string_stream.ts
